feat: dnd support
