+++ app/src/routes/zip/MeshSelectStepSidePanel.svelte
@@ -7,6 +7,8 @@
 	import { abbreviatePath } from '$lib/utils';
 	import { getTypeLabel, type MeshcodeData } from './utils';
 	import SecondaryButton from './SecondaryButton.svelte';
+	import { onMount } from 'svelte';
+	import { getCurrentWebview } from '@tauri-apps/api/webview';
 
 	type Props = {
 		meshcodeData: MeshcodeData | null;
@@ -27,6 +29,40 @@
 
 	let loading: boolean = $state(false);
 	let error: string | null = $state(null);
+
+	let isDnDEnter = $state(false);
+	let timerId = $state<number | null>(null);
+
+	onMount(() => {
+		const unlisten = getCurrentWebview().onDragDropEvent(async (event) => {
+			if (event.payload.type === 'enter') {
+				isDnDEnter = true;
+			}
+			if (event.payload.type === 'leave') {
+				isDnDEnter = false;
+			}
+			if (event.payload.type === 'over') {
+				isDnDEnter = true;
+				if (timerId) {
+					clearTimeout(timerId);
+				}
+				timerId = setTimeout(() => (isDnDEnter = false), 100);
+			}
+			if (event.payload.type === 'drop') {
+				isDnDEnter = false;
+				inputPaths = event.payload.paths.filter((path) => path.endsWith('.zip'));
+				if (inputPaths.length === 0) {
+					console.warn('No valid ZIP files dropped');
+					return;
+				}
+				await getMeshcodeWithPrefix(inputPaths);
+			}
+		});
+
+		return () => {
+			unlisten.then((f) => f());
+		};
+	});
 
 	async function openFileDialog() {
 		const res = await dialog.open({
@@ -73,6 +109,18 @@
 		}
 	}
 </script>
+
+{#if isDnDEnter}
+	<div
+		class="fixed inset-0 z-30 flex h-screen items-center justify-center bg-black/70 backdrop-blur-[2px]"
+	>
+		<div class="rounded-lg border-2 border-accent1 bg-gray-50 p-8 text-center shadow-lg">
+			<Icon icon="material-symbols:upload-file" class="mx-auto mb-4 text-6xl text-accent1" />
+			<p class="text-xl font-semibold text-base">PLATEAU の ZIP ファイルをドロップしてください</p>
+			<p class="mt-2 text-sm text-gray-600">複数ファイルにも対応しています</p>
+		</div>
+	</div>
+{/if}
 
 <div class="flex min-h-0 flex-1 flex-col gap-4">
 	<div class="flex flex-col gap-4">
